# backend/docs/swagger.yaml
openapi: "3.0.3"
info:
  title: SiAbsen API
  description: |
    REST API untuk **Sistem Absensi Karyawan (SiAbsen)** — Multi-Tenant SaaS.
    Semua endpoint memerlukan header `Authorization: Bearer <token>` kecuali
    `/api/v1/auth/login`. Header `X-Tenant-ID` wajib untuk identifikasi tenant.
  version: "1.0.0"
  contact:
    name: SiAbsen Support
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://api.siabsen.id
    description: Production
...
paths:
...
  /api/v1/absensi/checkin:
    post:
      tags: [Absensi]
      summary: Check-in karyawan
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [latitude, longitude]
              properties:
                latitude:
                  type: number
                longitude:
                  type: number
                foto_url:
                  type: string
      responses:
        "200":
          description: Check-in berhasil
        "400":
          description: Di luar area / sudah check-in
  /api/v1/absensi/checkout:
    post:
      tags: [Absensi]
      summary: Check-out karyawan
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [latitude, longitude]
              properties:
                latitude:
                  type: number
                longitude:
                  type: number
      responses:
        "200":
          description: Check-out berhasil
...
  /api/v1/izin:
    get:
      tags: [Izin]
      summary: List izin (semua / milik sendiri)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Daftar izin
    post:
      tags: [Izin]
      summary: Ajukan izin baru
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [jenis_izin, tanggal_mulai, tanggal_selesai, keterangan]
              properties:
                jenis_izin:
                  type: string
                  enum: [sakit, cuti, izin]
                tanggal_mulai:
                  type: string
                  format: date
                tanggal_selesai:
                  type: string
                  format: date
                keterangan:
                  type: string
      responses:
        "201":
          description: Izin diajukan
...
  /api/v1/users:
    get:
      tags: [Users]
      summary: Daftar semua user (admin)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List users

  # ── Profile ──────────────────────────────────────────────────────────────
  /api/v1/profile:
    get:
      tags: [Profile]
      summary: Ambil profil user saat ini
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Data profil

  # ── Superadmin ───────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...